Bugfix: numpy float32 object not JSON Serializable (#240)
* Bugfix: float32 object not JSON Serializable for `deepsparse.transformers.run_inference`

* Move `fix_numpy_types` functions from server.utils to deepsparse.transformers.helpers module

* Update usages accordingly

* Bugfix: Address review Comments

* Removed `numpy` alias
* Added a comment before wrapping `pipeline_object` in `process_dataset`
