+++ src/deepsparse/transformers/__init__.py
@@ -41,5 +41,6 @@ _check_transformers_install()
 
 
 from .helpers import *
+from .loaders import *
 from .pipelines import *
 from .server import *
